Implemented custom window decorator
Implemented a generic window decorator class to be merged into Hyena.Gui
that takes care of resizing and moving the window with no system window
manager decorations. Sets the proper cursor, etc.

Adds a subclass that does Cubano specific decoration rendering, and adapt
the CubanoWindow with this new decorator class.