Implemented new awesome GtkWindow subclasses
ShapedPopupWindow is a base class for easily implementing shaped
and composited/RGBA windows. It handles the shaping in composited
or uncomposited environments, and implements GtkContainer on its
own to support child margins so contents can be easily fitted to
the window shape.

BubbleWindow is a ShapedPopupWindow that draws a rounded window
frame with an arrow pointing to something, and adjusts its
margins automatically based on the arrow size and corner radius.