[ArtworkDisplay] Effectively complete and stable
Huge amount of progress here. Fully refactored the now playing
clutter UI into proper Actor subclasses. This fixes a lot of
various stability bugs.

Lots of performance improvements here as a result of the actual
NowPlayingInterface getting cleaned up an refactored -- we keep
the GtkClutterEmbed around and happy, same with the actual
now playing stage.

The actual ArtworkDisplay actor received most of the attention
here, and is very efficient. It uses an LRU cache for storing
and accessing artwork textures. A grid of ClutterClone actors
is constructed that is large enough to cover the entire screen.
When the ArtworkDisplay is allocated, the grid is populated from
the artwork cache.